Refactor math tests + select vector lib backend via kernel option (#136)
I've combined `test_libdevice` and `test_libmvec` into one `test_math`
suite. This removes a great deal of code duplication. (While at it, I
also sorted the test names in our github workflow file.)

Moreover, I've changed how we select the vector library -- instead of
using environment variables, we now can specify it via the kernel
kwargs. Environment variables are generally more awkward to work with.

Finally, I've made sleef the default vector library. It supports more
functions and architectures, and since we now vendor it, we don't have
to worry about portability issues.
