--- python/test/unit/cpu/test_libdevice.py
@@ -1,64 +1,0 @@
-import os
-import pytest
-import torch
-
-import triton
-import triton.language as tl
-from triton.language.extra import libdevice
-
-torch.manual_seed(0)
-
-
-def is_interpreter():
-    return os.environ.get('TRITON_INTERPRET', '0') == '1'
-
-
-def is_cpu():
-    return not is_interpreter() and \
-        triton.runtime.driver.active.get_current_target().backend == "cpu"
-
-
-float_dtypes = ['bfloat16', 'float16', 'float32', 'float64']
-
-
-@pytest.mark.parametrize("dtype_str", float_dtypes)
-@pytest.mark.parametrize("math_fn", [
-    "acos", "acosh", "asin", "asinh", "atan", "atanh", "cbrt", "cos", "cosh", "erf", "exp", "exp2", "expm1", "floor",
-    "isnan", "isinf", "log", "log1p", "log2", "log10", "rsqrt", "signbit", "sin", "sinh", "sqrt", "tan", "tanh"
-])
-@pytest.mark.parametrize("size", [1, 4, 16, 64])
-def test_libdevice(dtype_str, math_fn, size, device):
-    if not is_cpu():
-        pytest.skip("This test is CPU-specific")
-
-    @triton.jit
-    def kernel(src, dst, MATH_FN: tl.constexpr, BLOCK_SIZE: tl.constexpr):
-        idxs = tl.arange(0, BLOCK_SIZE)
-        x = tl.load(src + idxs)
-        y = getattr(libdevice, MATH_FN)(x)
-        tl.store(dst + idxs, y)
-
-    src = torch.rand((size, ), dtype=getattr(torch, dtype_str), device=device)
-
-    if math_fn == "acosh":
-        src = src.abs() + 1
-    if math_fn == "isnan" or math_fn == "isinf":
-        indices = torch.randint(low=0, high=size, size=(size // 2, ), device=device)
-        for i in indices:
-            if math_fn == "isnan":
-                src[i] = float("nan")
-            else:
-                src[i] = float(("+" if i % 2 else "-") + "inf")
-
-    if math_fn in ["isnan", "isinf", "signbit"]:
-        out_dtype = torch.bool
-    else:
-        out_dtype = getattr(torch, dtype_str)
-
-    res = torch.empty(src.shape, dtype=out_dtype, device=device)
-    kernel[(1, )](src, res, MATH_FN=math_fn, BLOCK_SIZE=size)
-    if math_fn == "cbrt":
-        ref = src.pow(1 / 3)
-    else:
-        ref = getattr(src, math_fn)()
-    torch.testing.assert_close(ref, res)
